refactor: new structure in the backend
