+++ backend/src/legacy/dto/AuthorDTO.ts
