+++ backend/src/legacy/dto/CategoryDTO.ts
