+++ backend/src/routes/TestRoutes.ts
@@ -1,0 +1,12 @@
+import {Router} from 'express';
+
+const router: Router = Router();
+
+router.get(
+  '/test',
+  (request, response): void => {
+    response.json({data: 'Hello, World!'});
+  },
+);
+
+export default router;
